# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LoL Tracker API",
    "description": "Backend del dashboard de rendimiento en ranked. Sustituye al monolito Streamlit.",
    "version": "2.0.0-dev"
  },
  "paths": {
...
    "/api/stats/summary": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Summary",
        "description": "Promedios como n\u00fameros. El monolito devolv\u00eda la KDA como el string \"5.0 / 2.0 / 10.0\",\nque la UI re-parseaba con `.split('/')` \u2014 origen del bug de `app.py:128`.",
        "operationId": "summary_api_stats_summary_get",
        "parameters": [
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/champions": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Champions",
        "description": "Incluye `winrate` y `kda_ratio` calculados en SQL: las dos columnas que la Tab 3 de\nStreamlit ped\u00eda y que la query nunca devolv\u00eda.",
        "operationId": "champions_api_stats_champions_get",
        "parameters": [
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChampionStats"
                  },
                  "title": "Response Champions Api Stats Champions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/heatmap": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Heatmap",
        "operationId": "heatmap_api_stats_heatmap_get",
        "parameters": [
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeatmapCell"
                  },
                  "title": "Response Heatmap Api Stats Heatmap Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/lp-trend": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Lp Trend",
        "description": "LP acumulado en orden cronol\u00f3gico. `has_lp` distingue \"0 LP\" de \"a\u00fan sin registrar\".\n\n`queue` acota la cola (420 = Solo/Duo): las normales sin LP diluyen la curva si cuentan como 0.",
        "operationId": "lp_trend_api_stats_lp_trend_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Queue"
            }
          },
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Lp Trend Api Stats Lp Trend Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/trends": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Kpi Trends",
        "description": "Serie temporal de KPIs de mejora (CS/min, DPM, KDA) de las \u00faltimas partidas v\u00e1lidas.\n\nIgnora remakes (< 5 min); los DPM de filas legacy sin participants llegan como 0.",
        "operationId": "kpi_trends_api_stats_trends_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrendPoint"
                  },
                  "title": "Response Kpi Trends Api Stats Trends Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/weekly": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Weekly Report",
        "description": "Resumen de la \u00faltima semana (7 d\u00edas, fecha UTC): partidas, winrate, KDA, top campe\u00f3n\ny mejor partida por rating. `most_played`/`best_match` son null si no hay partidas en la ventana.",
        "operationId": "weekly_report_api_stats_weekly_get",
        "parameters": [
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/matchups/{user_champion}/{enemy_champion}": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Matchup Stats",
        "description": "Estad\u00edsticas hist\u00f3ricas del cruce de dos campeones.\n\nFiltra por campe\u00f3n del usuario contra el rival de l\u00ednea. Con cero partidas devuelve\ntodo a 0 (no 404), porque la vista Matchups sigue funcionando con notas aunque no\nhaya historial del cruce.",
        "operationId": "matchup_stats_api_stats_matchups__user_champion___enemy_champion__get",
        "parameters": [
          {
            "name": "user_champion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Champion"
            }
          },
          {
            "name": "enemy_champion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Enemy Champion"
            }
          },
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchupStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stats/constitution": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Constitution Status",
        "operationId": "constitution_status_api_stats_constitution_get",
        "parameters": [
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstitutionStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matchup-notes/{user_champion}/{enemy_champion}": {
      "get": {
        "tags": [
          "matchup-notes"
        ],
        "summary": "Get Notes",
        "description": "Notas del cruce. Si nunca se guard\u00f3 nada, devuelve notas vac\u00edas (no 404).",
        "operationId": "get_notes_api_matchup_notes__user_champion___enemy_champion__get",
        "parameters": [
          {
            "name": "user_champion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Champion"
            }
          },
          {
            "name": "enemy_champion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Enemy Champion"
            }
          },
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchupNotes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "matchup-notes"
        ],
        "summary": "Put Notes",
        "description": "Guarda/reemplaza las notas del cruce (PUT semantics).",
        "operationId": "put_notes_api_matchup_notes__user_champion___enemy_champion__put",
        "parameters": [
          {
            "name": "user_champion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Champion"
            }
          },
          {
            "name": "enemy_champion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Enemy Champion"
            }
          },
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchupNotesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchupNotes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChampionStats": {
        "properties": {
          "champion": {
            "type": "string",
            "title": "Champion"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "winrate": {
            "type": "number",
            "title": "Winrate"
          },
          "avg_kills": {
            "type": "number",
            "title": "Avg Kills"
          },
          "avg_deaths": {
            "type": "number",
            "title": "Avg Deaths"
          },
          "avg_assists": {
            "type": "number",
            "title": "Avg Assists"
          },
          "kda_ratio": {
            "type": "number",
            "title": "Kda Ratio"
          },
          "avg_cs_min": {
            "type": "number",
            "title": "Avg Cs Min"
          },
          "avg_dpm": {
            "type": "number",
            "title": "Avg Dpm",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "champion",
          "games_played",
          "wins",
          "winrate",
          "avg_kills",
          "avg_deaths",
          "avg_assists",
          "kda_ratio",
          "avg_cs_min"
        ],
        "title": "ChampionStats",
        "description": "Incluye `winrate` y `kda_ratio`, que la query original nunca devolv\u00eda pese a que la\nTab 3 las ped\u00eda -> KeyError permanente. Ahora se calculan en SQL."
      },
...
      "ConstitutionStatus": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "STOP",
              "ON_FIRE",
              "NEUTRAL",
              "NO_DATA"
            ],
            "title": "State"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "loss_streak": {
            "type": "integer",
            "title": "Loss Streak"
          },
          "last_results": {
            "items": {
              "type": "boolean"
            },
            "type": "array",
            "title": "Last Results",
            "description": "M\u00e1s reciente primero; True = victoria"
          },
          "last_champion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Champion",
            "description": "Campe\u00f3n de la \u00faltima partida"
          },
          "off_pool": {
            "type": "boolean",
            "title": "Off Pool",
            "description": "La \u00faltima partida fue fuera del champion pool",
            "default": false
          },
          "champion_pool": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Champion Pool",
            "description": "Pool declarado, para contexto"
          }
        },
        "type": "object",
        "required": [
          "state",
          "message",
          "loss_streak",
          "last_results"
        ],
        "title": "ConstitutionStatus",
        "description": "La regla anti-tilt, movida del sidebar de Streamlit al backend.\n\nIncluye la comprobaci\u00f3n de champion pool contra la configuraci\u00f3n **persistida**, de modo que\nla alerta de \"has jugado algo fuera de tus mains\" ya no depende de un `text_area` vol\u00e1til."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HeatmapCell": {
        "properties": {
          "day_of_week": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Day Of Week",
            "description": "0 = Domingo, 1 = Lunes ... 6 = S\u00e1bado"
          },
          "time_block": {
            "type": "string",
            "title": "Time Block",
            "description": "Madrugada | Ma\u00f1ana | Tarde | Noche"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "losses": {
            "type": "integer",
            "title": "Losses"
          },
          "winrate": {
            "type": "number",
            "title": "Winrate"
          }
        },
        "type": "object",
        "required": [
          "day_of_week",
          "time_block",
          "games_played",
          "wins",
          "losses",
          "winrate"
        ],
        "title": "HeatmapCell"
      },
...
      "MatchupNotes": {
        "properties": {
          "user_champion": {
            "type": "string",
            "title": "User Champion"
          },
          "enemy_champion": {
            "type": "string",
            "title": "Enemy Champion"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_champion",
          "enemy_champion"
        ],
        "title": "MatchupNotes",
        "description": "Notas persistidas para un emparejamiento concreto."
      },
      "MatchupNotesUpdate": {
        "properties": {
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "title": "MatchupNotesUpdate",
        "description": "Payload del PUT: reemplazo completo de las notas del cruce."
      },
      "MatchupStats": {
        "properties": {
          "user_champion": {
            "type": "string",
            "title": "User Champion"
          },
          "enemy_champion": {
            "type": "string",
            "title": "Enemy Champion"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played",
            "default": 0
          },
          "wins": {
            "type": "integer",
            "title": "Wins",
            "default": 0
          },
          "losses": {
            "type": "integer",
            "title": "Losses",
            "default": 0
          },
          "winrate": {
            "type": "number",
            "title": "Winrate",
            "default": 0.0
          },
          "avg_kills": {
            "type": "number",
            "title": "Avg Kills",
            "default": 0.0
          },
          "avg_deaths": {
            "type": "number",
            "title": "Avg Deaths",
            "default": 0.0
          },
          "avg_assists": {
            "type": "number",
            "title": "Avg Assists",
            "default": 0.0
          },
          "kda_ratio": {
            "type": "number",
            "title": "Kda Ratio",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "user_champion",
          "enemy_champion"
        ],
        "title": "MatchupStats",
        "description": "Estad\u00edsticas hist\u00f3ricas del cruce de dos campeones (T\u00fa contra enemigo)."
      },
...
      "StatsSummary": {
        "properties": {
          "total_games": {
            "type": "integer",
            "title": "Total Games"
          },
          "total_wins": {
            "type": "integer",
            "title": "Total Wins"
          },
          "winrate": {
            "type": "number",
            "title": "Winrate"
          },
          "avg_kills": {
            "type": "number",
            "title": "Avg Kills"
          },
          "avg_deaths": {
            "type": "number",
            "title": "Avg Deaths"
          },
          "avg_assists": {
            "type": "number",
            "title": "Avg Assists"
          },
          "kda_ratio": {
            "type": "number",
            "title": "Kda Ratio"
          },
          "avg_cs_min": {
            "type": "number",
            "title": "Avg Cs Min"
          }
        },
        "type": "object",
        "required": [
          "total_games",
          "total_wins",
          "winrate",
          "avg_kills",
          "avg_deaths",
          "avg_assists",
          "kda_ratio",
          "avg_cs_min"
        ],
        "title": "StatsSummary"
      },
...
      "TrendPoint": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "cs_min": {
            "type": "number",
            "title": "Cs Min"
          },
          "dpm": {
            "type": "number",
            "title": "Dpm"
          },
          "kda": {
            "type": "number",
            "title": "Kda"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "timestamp",
          "cs_min",
          "dpm",
          "kda"
        ],
        "title": "TrendPoint",
        "description": "Un punto de la serie temporal de KPIs de mejora (\u00faltimas N partidas v\u00e1lidas).\n\nOrden cronol\u00f3gico ascendente (partida m\u00e1s antigua primero) para que las gr\u00e1ficas de\nl\u00ednea recorran el tiempo de izquierda a derecha. `timestamp` es la hora de juego en UTC."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeeklyReport": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "total_games": {
            "type": "integer",
            "title": "Total Games"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "losses": {
            "type": "integer",
            "title": "Losses"
          },
          "winrate": {
            "type": "number",
            "title": "Winrate",
            "default": 0.0
          },
          "avg_kda": {
            "type": "number",
            "title": "Avg Kda",
            "default": 0.0
          },
          "most_played": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeeklyTopChampion"
              },
              {
                "type": "null"
              }
            ]
          },
          "best_match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeeklyBestMatch"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end",
          "total_games",
          "wins",
          "losses"
        ],
        "title": "WeeklyReport",
        "description": "Resumen de los \u00faltimos 7 d\u00edas (ventana seg\u00fan fecha UTC) para el Reporte Semanal."
      },
      "WeeklyTopChampion": {
        "properties": {
          "champion": {
            "type": "string",
            "title": "Champion"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          }
        },
        "type": "object",
        "required": [
          "champion",
          "games",
          "wins"
        ],
        "title": "WeeklyTopChampion",
        "description": "Campe\u00f3n m\u00e1s jugado de la semana (nombre visible; el avatar lo resuelve el frontend v\u00eda Data Dragon)."
      }
    }
  }
}